depositoweb/requirements.txt
Flask
flask-mysqldb
gunicorn
python-telegram-bot
Flask-SQLAlchemy
pymysql